+++ tools/frb_internal/lib/src/frb_example_pure_dart_generator/dart/direct_sources.dart
@@ -1,6 +1,6 @@
+import 'package:flutter_rust_bridge_internal/src/frb_example_pure_dart_generator/basic_type_infos.dart';
 import 'package:flutter_rust_bridge_internal/src/frb_example_pure_dart_generator/dart/benchmark.dart';
 import 'package:flutter_rust_bridge_internal/src/frb_example_pure_dart_generator/dart/builder.dart';
-import 'package:flutter_rust_bridge_internal/src/frb_example_pure_dart_generator/rust/direct_sources.dart';
 import 'package:recase/recase.dart';
 
 Map<String, String> generateDartDirectSources() {
